[refactor]: schema.js 모듈 분리
